+++ src/ingest_validation_tools/submission.py
@@ -1,7 +1,8 @@
-import csv
+from csv import DictReader
 from pathlib import Path
 from datetime import datetime
 import re
+from collections import defaultdict
 
 from ingest_validation_tools.validation_utils import (
     get_metadata_tsv_errors,
@@ -76,7 +77,7 @@ class Submission:
     def _get_single_tsv_external_errors(self, type, path):
         errors = {}
         with open(path) as f:
-            rows = list(csv.DictReader(f, dialect='excel-tab'))
+            rows = list(DictReader(f, dialect='excel-tab'))
             if not rows:
                 errors['Warning'] = f'File has no data rows.'
             if self.directory_path:
@@ -85,7 +86,7 @@ class Submission:
                     data_dir_errors = self._get_data_dir_errors(
                         type, full_data_path)
                     if data_dir_errors:
-                        errors[f'{path.name} (row {i+1})'] = data_dir_errors
+                        errors[f'{path.name} (row {i+2})'] = data_dir_errors
         return errors
 
     def _get_data_dir_errors(self, type, path):
@@ -102,11 +103,28 @@ class Submission:
         return errors
 
     def _get_no_ref_errors(self):
-        errors = {}
-        # TODO
-        return errors
+        referenced_data_paths = set(self._get_data_references().keys())
+        non_metadata_paths = {
+            path.name for path in self.directory_path.iterdir()
+            if not path.name.endswith('-metadata.tsv')
+        }
+        unreferenced_paths = non_metadata_paths - referenced_data_paths
+        return [str(path) for path in unreferenced_paths]
 
     def _get_multi_ref_errors(self):
         errors = {}
-        # TODO
+        data_references = self._get_data_references()
+        for path, references in data_references.items():
+            if len(references) > 1:
+                errors[path] = references
         return errors
+
+    def _get_data_references(self):
+        # TODO: Move this to __init__
+        data_references = defaultdict(list)
+        for tsv_path in self.effective_tsv_paths.values():
+            with open(tsv_path) as f:
+                for i, row in enumerate(DictReader(f, dialect='excel-tab')):
+                    reference = f'{tsv_path} (row {i+2})'
+                    data_references[row['data_path']].append(reference)
+        return data_references