+++ test-examples.sh
@@ -14,7 +14,7 @@ for EXAMPLE in examples/*; do
 done
 
 for GOOD in examples/good-*/README.md; do
-  grep 'No errors!' $GOOD || die "$GOOD should not be an error report."
+  grep 'No errors!' $GOOD > /dev/null || die "$GOOD should not be an error report."
 done
 
 for BAD in examples/bad-*/README.md; do